// src/types/onyx/Transaction.ts
type ReceiptError = {
    /** Path of the receipt file */
    source: string;

    /** Name of the receipt file */
    filename: string;

    /** Action that caused the error */
    action: string;

    /** Parameters required to retry the failed action */
    retryParams: StartSplitBilActionParams | CreateTrackExpenseParams | RequestMoneyInformation | ReplaceReceipt;

    /** Transaction associated with the failed receipt flow */
    transactionID?: string;

    /** The type of receipt error */
    error: typeof CONST.IOU.RECEIPT_ERROR;
};